New pragma infrastructure
