+++ tests/insert-pragma/graphql/jsfmt.spec.js
@@ -1,0 +1,1 @@
+run_spec(__dirname, ["graphql"], { insertPragma: true });
