+++ tests/insert-pragma/graphql/with-pragma.graphql
@@ -1,0 +1,5 @@
+# @format
+
+type Foo {
+  one: Type
+}
