+++ tests/insert-pragma/graphql/without-pragma.graphql
@@ -1,0 +1,3 @@
+type Foo {
+  one: Type
+}
