+++ tests/insert-pragma/js/module-with-late-docblock.js
