+++ tests/insert-pragma/js/module-with-pragma.js
