+++ tests/insert-pragma/js/module-without-docblock.js
