+++ tests/insert-pragma/js/module-without-pragma.js
