+++ tests/insert-pragma/js/trailing-spaces-first-line.js
