--- tests/insert-pragma/jsfmt.spec.js
@@ -1,1 +1,0 @@
-run_spec(__dirname, ["flow"], { insertPragma: true });
