+++ tests/require-pragma/graphql/with-pragma.graphql
@@ -1,0 +1,6 @@
+# @format
+
+type    Foo
+    {
+          one:    Type
+      }
