+++ tests/require-pragma/graphql/without-pragma.graphql
@@ -1,0 +1,4 @@
+type    Foo
+    {
+          one:    Type
+      }
