+++ tests/require-pragma/js/jsfmt.spec.js
@@ -1,0 +1,1 @@
+run_spec(__dirname, ["flow", "babylon", "typescript"], { requirePragma: true });
