+++ tests/require-pragma/js/module-with-pragma.js
