+++ tests/require-pragma/js/module-without-pragma.js
