--- tests/require-pragma/jsfmt.spec.js
@@ -1,1 +1,0 @@
-run_spec(__dirname, ["flow"], { requirePragma: true });
